feat(projects): add project management features including creation, deletion, and filtering feat(projects): implement chat interface for user interaction and project discussions feat(projects): create code explorer and preview panel for project files and outputs feat(projects): add loading skeletons for better user experience during data fetching feat(projects): implement responsive design for project management components fix(projects): update import paths for project components to reflect new structure chore(projects): refactor project-related components for improved maintainability and readability
feat(ai): implement AI generation functionality with support for multiple models and message types
chore(api): rename ai-sdk to ai for consistency and clarity in imports
feat(pipelines): add RooCodePipeline implementation for enhanced project processing
fix(prompts): update prompts to use new ai module for message types
feat(preview): introduce DockerRunner for managing project previews in Docker containers
chore(preview): refactor import paths for consistency across preview modules
feat(k8s): add K8sRunner class as a placeholder for future Kubernetes integration
style: standardize file naming conventions for better readability and maintainability
