+++ app/(logged-in)/billing/components/cancel-subscription-button.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { Button } from '@/components/ui/button';
-import { ConfirmationDialog } from './ConfirmationDialog';
+import { ConfirmationDialog } from './confirmation-dialog';
 
 export function CancelSubscriptionButton() {
   const [showConfirmation, setShowConfirmation] = useState(false);
