+++ app/(logged-in)/projects/[id]/components/AssistantActionsCard.tsx
