+++ app/(logged-in)/projects/[id]/components/BuildingStatus.tsx
