+++ app/(logged-in)/projects/[id]/components/ChatInput.tsx
