+++ app/(logged-in)/projects/[id]/components/CodeExplorer.tsx
