+++ app/(logged-in)/projects/[id]/components/FileTree.tsx
