+++ app/(logged-in)/projects/[id]/components/ModelBanner.tsx
