+++ app/(logged-in)/projects/components/ProjectCardSkeleton.tsx
@@ -1,22 +1,6 @@
 import { Skeleton } from "@/components/ui/skeleton";
 import { Card, CardHeader, CardContent } from "@/components/ui/card";
 
-export function ProjectsLoadingSkeleton() {
-  return (
-    <div className="container mx-auto py-8">
-      <div className="flex justify-between items-center mb-8">
-        <Skeleton className="h-8 w-48" />
-        <Skeleton className="h-10 w-32" />
-      </div>
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-        {Array.from({ length: 6 }).map((_, i) => (
-          <ProjectCardSkeleton key={i} />
-        ))}
-      </div>
-    </div>
-  );
-}
-
 export function ProjectCardSkeleton() {
   return (
     <div className="block group">
@@ -38,22 +22,6 @@ export function ProjectCardSkeleton() {
         </CardContent>
         <div className="absolute inset-0 bg-gradient-to-b from-transparent to-muted/10 opacity-0 group-hover:opacity-100 transition-opacity pointer-events-none" />
       </Card>
-    </div>
-  );
-}
-
-export function ProjectModalSkeleton() {
-  return (
-    <div className="space-y-6 p-6">
-      <Skeleton className="h-7 w-48" />
-      <div className="space-y-4">
-        <Skeleton className="h-10 w-full" />
-        <Skeleton className="h-24 w-full" />
-      </div>
-      <div className="flex justify-end space-x-2">
-        <Skeleton className="h-10 w-24" />
-        <Skeleton className="h-10 w-24" />
-      </div>
     </div>
   );
 }
