+++ lib/llm/api/index.ts
@@ -1,1 +1,1 @@
-export * from './ai-sdk';
+export * from './ai';
