+++ lib/llm/pipelines/rooCode.ts
