+++ lib/llm/prompts/index.ts
@@ -1,2 +1,2 @@
 export * from './naive';
-export * from './roo-code';
+export * from './rooCode';
